introduce-invariant-checker-move: Move invariant checks to SchedulerInvariantChecker and retire runtime_checker mixin
Mechanical cut + paste for the ``introduce-invariant-checker`` mech move
(tail commit of the ``SchedulerRuntimeCheckerMixin`` split).

Cut the invariant-check methods (the @staticmethod-converted ones from
prep plus the already-static ``_check_pool_invariant``) from
``SchedulerRuntimeCheckerMixin`` and paste them into the
``SchedulerInvariantChecker`` class body in
``scheduler_components/invariant_checker.py``. Drop ``@staticmethod``
decorators; simplify ``self: "SchedulerInvariantChecker"`` annotation to
bare ``self`` (in class context the type is implicit). Strip the
``SchedulerRuntimeCheckerMixin.<method>(self, ...)`` qualified form on
internal sibling calls to plain ``self.<method>(...)`` (the methods are
now in the same class). Method bodies otherwise byte-identical.

All callers updated:
  ``self.<method>(self.invariant_checker, ...)`` →
  ``self.invariant_checker.<method>(...)``
(pure prefix transformation). The callsites in ``scheduler.py``
(``on_idle`` / ``_maybe_log_idle_metrics`` / ``run_batch`` /
``event_loop_overlap``) plus the callsite in ``create_scheduler_watchdog``
(also in ``scheduler.py`` post-pre-prep) are rewritten.

Final cleanup:
- Drop ``SchedulerRuntimeCheckerMixin`` from ``Scheduler``'s inheritance
  list and from the import in ``scheduler.py``.
- Delete ``python/sglang/srt/managers/scheduler_runtime_checker_mixin.py``
  (now empty after the invariant-check method bodies are moved out and
  ``create_scheduler_watchdog`` already moved in pre-prep).

Refactor chain ID: introduce-invariant-checker-move
